커밋

+++ WGNProject/src/main/java/com/smhrd/web/entity/t_review.java
@@ -28,7 +28,7 @@ public class t_review {
  private String img_link;
 
  // 별점 
- private Integer ratings;
+ private float ratings;
 
  // 좋아요 수 
  private Integer likes;
